Update version and add MgSoftDev.PrismPlus.Returning project
- Updated version of MgSoftDev.PrismPlus from undefined to 1.0.2
- Added new project MgSoftDev.PrismPlus.Returning targeting various frameworks
- Refactored AsyncCommand.xaml.cs to use new ReturningCommand
- Added new Command classes in MgSoftDev.PrismPlus.Returning/Commands/
- Included reference to MgSoftDev.PrismPlus.Returning in Samples project

The main purpose of these changes is to implement DelegateCommands types that extend the functionality of the Prism Library by providing support for returning a result. Updating the version of the MgSoftDev.PrismPlus project to ensure compatibility with the newly added MgSoftDev.PrismPlus.Returning project and its features. In addition, the Samples project has been updated with a new reference to better demonstrate the new capabilities of our library.

Took 2 hours 43 minutes