Update version numbers and add global.json file
Updated the version numbers in the .csproj files for MgSoftDev.PrismPlus and MgSoftDev.PrismPlus.Returning projects from 1.0.4 to 1.0.5. Also, the version of the MgSoftDev.Returning package reference was upgraded from 1.0.9 to 1.1.0 in both projects and in the Samples project. In addition, a new global.json file was added.

The version numbers were updated to reflect new enhancements and bug fixes in the code. The MgSoftDev.Returning reference version was upgraded to use new features from that package. A global.json file was added to specify a required .NET SDK version and rolling forward policy.

Took 11 minutes